+++ src/sentry/issues/grouptype.py
@@ -22,7 +22,7 @@ if TYPE_CHECKING:
     from sentry.models.organization import Organization
     from sentry.models.project import Project
     from sentry.users.models.user import User
-    from sentry.workflow_engine.processors.detector import DetectorHandler
+    from sentry.workflow_engine.models.detector import DetectorHandler
 import logging
 
 logger = logging.getLogger(__name__)
